Added the check_saved_games and check_json functions.
The check_saved_games function checks if the saved_games folder exists, if not, it creates one and sends the user back to the main menu. It also checks if there are any files in the saved_games folder if it does exist. If there aren't any files, the user is sent back to the main menu. It also checks if there is at least 1 file in the saved_games folder with the .json extension. If there isn't any, the user will be sent back to the main menu.